Fix the Facebook native
