+++ src/facebook/web/index.html
@@ -1,9 +1,16 @@
 <script src="https://connect.facebook.net/en_US/sdk/xfbml.ad.js#xfbml=1&version=v2.5&appId=180444840287" async></script>
-<div class="fb-ad ad-slot--facebook"
+<div class="fb-ad"
     data-placementid="180444840287_10154600557405288"
     data-format="native"
     data-nativeadid="fb_ad_root_mpu"></div>
 <aside class="creative creative--facebook" id="fb_ad_root_mpu">
+    <div class="roundel has-popup">
+        <button class="roundel__button u-button-reset js-toggle" aria-controls="popup" aria-expanded="true">Ad {{#svg}}arrow-down{{/svg}}</button>
+        <div id="popup" class="popup">
+            <p>Advertising on the Guardian is created and paid for by third parties.</p>
+            <a href="%%CLICK_URL_UNESC%%https://www.theguardian.com/advertising-on-the-guardian" data-link-name="about">How advertising supports the Guardian {{#svg}}arrow-right{{/svg}}</a>
+        </div>
+    </div>
     <a class="fbAdLink advert advert--facebook">
         <div class="fbAdMedia advert__image-container"></div>
         <h2 class="fbAdTitle advert__title"></h2>
